+++ include/engine/api/route_parameters.hpp
@@ -76,8 +76,7 @@ struct RouteParameters : public BaseParameters
         Weight = 0x08,
         Datasources = 0x10,
         Speed = 0x20,
-        DatasourceNames = 0x40,
-        All = Duration | Nodes | Distance | Weight | Datasources | Speed | DatasourceNames
+        All = Duration | Nodes | Distance | Weight | Datasources | Speed
     };
 
     RouteParameters() = default;